enhancement: refactor factories structure in core package
