+++ packages/core/core/src/factories/constants.ts
@@ -1,0 +1,3 @@
+export const symbols = {
+  CustomController: Symbol('StrapiCustomCoreController'),
+} as const;
